Servicio EntidadPaciente y paciente, se esta configurando los TEST

+++ microservicio/dominio/src/main/java/com/ceiba/consultorio/modelo/entidad/Entidad.java
@@ -8,10 +8,10 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class Entidad {
-    private Long idEntidad;
+    private Integer idEntidad;
     private String eps;
 
-    public Entidad(Long idEntidad, String eps) {
+    public Entidad(Integer idEntidad, String eps) {
         this.idEntidad = idEntidad;
         this.eps = eps;
     }
